Rename headers to unique names
ToDo:
- Implement events
- Fix install routine
