+++ examples/cuda/build_kernel.cpp
@@ -7,7 +7,7 @@
 #include <hpx/include/iostreams.hpp>
 #include <hpx/lcos/future.hpp>
 
-#include "../../cuda.hpp"
+#include "../../hpx_cuda.hpp"
 
 #include <unistd.h>
