+++ examples/cuda/list_devices.cpp
@@ -7,7 +7,7 @@
 #include <hpx/include/iostreams.hpp>
 #include <hpx/lcos/future.hpp>
 
-#include "../../cuda.hpp"
+#include "../../hpx_cuda.hpp"
 #include <sstream>
 
 // hpx_main, is the actual main called by hpx
